Move genotyper_config into types.h
