+++ include/types.h
@@ -1,5 +1,4 @@
-#ifndef GLNEXUS_TYPES_H
-#define GLNEXUS_TYPES_H
+#pragma once
 
 #include <string>
 #include <vector>
@@ -343,6 +342,38 @@ struct StatsRangeQuery {
     }
 };
 
+enum class GLnexusOutputFormat {
+    /// Compressed bcf (default option)
+    BCF,
+
+    /// Uncompressed vcf (for ease of comparison in small cases)
+    VCF,
+};
+struct genotyper_config {
+    /// Require any allele call to be supported by at least this depth
+    size_t required_dp = 0;
+
+    /// FORMAT field to consult for per-allele depth in VCF records
+    std::string allele_dp_format = "AD";
+
+    /// The symbolic allele used in gVCF reference confidence models
+    std::string ref_symbolic_allele = "<NON_REF>";
+
+    /// FORMAT field to consult for reference depth in gVCF reference records
+    std::string ref_dp_format = "MIN_DP";
+
+    // Should the genotyper write a record describing each call loss?
+    // If true, the output is recorded in YAML format in the
+    // a file named [BCF/VCF output file].residuals.yml
+    bool output_residuals = false;
+
+    /// Output format (default = bcf), choices = "BCF", "VCF"
+    GLnexusOutputFormat output_format = GLnexusOutputFormat::BCF;
+
+    genotyper_config() = default;
+
+    genotyper_config(GLnexusOutputFormat _output_format) : output_format(_output_format) {}
+};
+
 } //namespace GLnexus
 
-#endif
